bitbucket_hook,functions: Improve output buffer handling.
Output buffers (if active) are explicitly flushed in case of fatal errors,
and output buffering is also enabled in case of backgrounding,
at the end discarding the buffered output.

The latter prevents random terminations of the PHP webhook script,
which died silently in case too much output was being flushed after the listener
was shooed away.